requirements.txt
dearpygui
pyperclip
requests
pillow
numpy
keyboard
dxcam[cv2]

# Optional
pyinstaller
black
pyinstaller_versionfile
customtkinter